% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediantest.R
\name{medianTest}
\alias{medianTest}
\title{Median test}
\usage{
medianTest(x, fac, n = 999, ...)
}
\arguments{
\item{x}{a vector of a quantitative variable}

\item{fac}{a factor defining 2 groups}

\item{n}{the number of permutations for the Monte Carlo procedure}

\item{...}{further arguments passed to as.randtest}
}
\description{
This function implements a non-parametric test of differences between medians in two samples.
}
\examples{
## look at the data
head(VL)

## make test
set.seed(1)
x <- medianTest(VL$titer, VL$type)

## look at results
x
plot(x)

}
\author{
T. Jombart
}

